fix E2E for cloud (#7710)
