+++ e2e/cypress/integration/enterprise/elasticsearch_autocomplete/renaming_spec.js
@@ -8,7 +8,7 @@
 // ***************************************************************
 
 // Stage: @prod
-// Group: @enterprise @elasticsearch @autocomplete
+// Group: @enterprise @elasticsearch @autocomplete @not_cloud
 
 import {getRandomId} from '../../../utils';
 
@@ -23,6 +23,8 @@ describe('Autocomplete with Elasticsearch - Renaming', () => {
     let testChannel;
 
     before(() => {
+        cy.shouldNotRunOnCloudEdition();
+
         // * Check if server has license for Elasticsearch
         cy.apiRequireLicenseForFeature('Elasticsearch');
