+++ e2e/cypress/integration/enterprise/elasticsearch_autocomplete/users_in_message_input_box_spec.js
@@ -7,7 +7,7 @@
 // - Use element ID when selecting an element. Create one if none.
 // ***************************************************************
 
-// Group: @enterprise @elasticsearch @autocomplete
+// Group: @enterprise @elasticsearch @autocomplete @not_cloud
 
 import * as TIMEOUTS from '../../../fixtures/timeouts';
 
@@ -23,6 +23,8 @@ describe('Autocomplete with Elasticsearch - Users', () => {
     let testTeam;
 
     before(() => {
+        cy.shouldNotRunOnCloudEdition();
+
         // * Check if server has license for Elasticsearch
         cy.apiRequireLicenseForFeature('Elasticsearch');
