+++ e2e/cypress/integration/enterprise/system_console/compliance/compliance_export_ui_spec.js
@@ -84,59 +84,6 @@ describe('Compliance Export', () => {
         });
     });
 
-    it('MM-T3439 - Download Compliance Export Files - S3 Bucket Storage', () => {
-        // # Go to file storage settings Page
-        cy.visit('/admin_console/environment/file_storage');
-        cy.get('.admin-console__header', {timeout: TIMEOUTS.ONE_MIN}).should('be.visible').and('have.text', 'File Storage');
-
-        const {
-            minioAccessKey,
-            minioSecretKey,
-            minioS3Bucket,
-            minioS3Endpoint,
-            minioS3SSL,
-        } = Cypress.env();
-
-        // # Update S3 Storage settings
-        cy.findByTestId('FileSettings.DriverNamedropdown').select('amazons3');
-        cy.findByTestId('FileSettings.AmazonS3Bucketinput').clear().type(minioS3Bucket);
-        cy.findByTestId('FileSettings.AmazonS3AccessKeyIdinput').clear().type(minioAccessKey);
-        cy.findByTestId('FileSettings.AmazonS3SecretAccessKeyinput').clear().type(minioSecretKey);
-        cy.findByTestId('FileSettings.AmazonS3Endpointinput').clear().type(minioS3Endpoint);
-        cy.findByTestId(`FileSettings.AmazonS3SSL${minioS3SSL}`).check();
-
-        // # Save file storage settings
-        cy.findByTestId('saveSetting').click();
-        waitUntilConfigSave();
-
-        // # Test connection and verify that it's successful
-        cy.findByRole('button', {name: 'Test Connection'}).click();
-        cy.findByText('Connection was successful').should('be.visible');
-
-        // # Go to compliance page and enable export
-        cy.uiGoToCompliancePage();
-        cy.uiEnableComplianceExport();
-        cy.uiExportCompliance();
-
-        // # Navigate to a team and post an attachment
-        gotoTeamAndPostImage();
-
-        // # Go to compliance page and start export
-        cy.uiGoToCompliancePage();
-        cy.uiExportCompliance();
-
-        // # Get the first row
-        cy.get('.job-table__table').find('tbody > tr').eq(0).as('firstRow');
-
-        // # Get the download link
-        cy.get('@firstRow').findByText('Download').parents('a').should('exist').then((fileAttachment) => {
-            const fileURL = fileAttachment.attr('href');
-
-            // * Download link should not exist this time
-            cy.apiDownloadFileAndVerifyContentType(fileURL);
-        });
-    });
-
     it('MM-T1168 - Compliance Export - Run Now, entry appears in job table', () => {
         // # Go to compliance page and enable export
         cy.uiGoToCompliancePage();
@@ -220,10 +167,3 @@ describe('Compliance Export', () => {
         cy.get('@firstRow').find('td:eq(1)').should('have.text', 'Canceled');
     });
 });
-
-// # Wait until the Saving text becomes Save
-const waitUntilConfigSave = () => {
-    cy.waitUntil(() => cy.get('#saveSetting').then((el) => {
-        return el[0].innerText === 'Save';
-    }));
-};
