+++ e2e/cypress/integration/enterprise/system_console/support_packet_generation_spec.js
@@ -10,38 +10,14 @@
 // Stage: @prod
 // Group: @enterprise @system_console
 
-// # Goes to the System Scheme page as System Admin
-const goToAdminConsole = () => {
-    cy.apiAdminLogin();
-    cy.visit('/admin_console');
-};
-
-const goToSupportPacketGenerationModal = () => {
-    cy.findByRole('button', {name: 'Menu Icon'}).should('exist').click();
-
-    cy.findByRole('button', {name: 'Commercial Support dialog'}).click();
-
-    // * Ensure the download support packet button exist and that text regarding setting the proper settings exist
-    cy.findByRole('link', {name: 'Download Support Packet'}).should('exist');
-};
-
 describe('Support Packet Generation', () => {
     before(() => {
-        // Running on E10/E20 License
         cy.apiRequireLicense();
-        cy.apiAdminLogin();
-
-        cy.apiUpdateConfig({
-            LogSettings: {
-                EnableFile: true,
-                FileLevel: 'ERROR',
-            },
-        });
     });
 
     it('MM-T3849 - Commercial Support Dialog UI - E10/E20 License', () => {
         // # Go to System Console
-        goToAdminConsole();
+        cy.visit('/admin_console');
 
         goToSupportPacketGenerationModal();
 
@@ -50,12 +26,23 @@ describe('Support Packet Generation', () => {
 
     it('MM-T3818 - Commercial Support Dialog UI - Links', () => {
         // # Go to System Console
-        goToAdminConsole();
+        cy.visit('/admin_console');
 
         goToSupportPacketGenerationModal();
 
-        // * Veryify the links exist that take you to loggin page and ticket page exist
+        // * Verify that the "submit a support ticket." link exist and points to Customer Support Request page
         cy.findByRole('link', {name: 'submit a support ticket.'}).should('have.attr', 'href').and('include', 'https://support.mattermost.com/hc/en-us/requests/new');
+
+        // * Verify that the "here" link exist and points to Logging admin page
         cy.findByRole('link', {name: 'here'}).should('have.attr', 'href').and('include', '/admin_console/environment/logging');
     });
 });
+
+const goToSupportPacketGenerationModal = () => {
+    // # Open system menu and click Customer Support
+    cy.findByRole('button', {name: 'Menu Icon'}).should('exist').click();
+    cy.findByRole('button', {name: 'Commercial Support dialog'}).click();
+
+    // * Ensure the download support packet button exist and that text regarding setting the proper settings exist
+    cy.findByRole('link', {name: 'Download Support Packet'}).should('exist');
+};
