+++ e2e/cypress/integration/enterprise/visual_regression/system_console/authentication_section_spec.js
@@ -73,9 +73,6 @@ describe('System Console - Authentication', () => {
     ];
 
     before(() => {
-        // * Check if server has license for feature
-        cy.apiRequireLicenseForFeature('Elasticsearch');
-
         // # Go to system admin then verify admin console URL and header
         cy.visit('/admin_console/about/license');
         cy.url().should('include', '/admin_console/about/license');
